% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rustr.R
\name{check_rustr}
\alias{check_rustr}
\title{Check rustinr status}
\usage{
check_rustr(detail = FALSE)
}
\arguments{
\item{detail}{print detail info, FALSE by default.}
}
\description{
\code{check_rustr} checks the status of rustinr, rustc, and cargo installation.
}

