% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rust}
\alias{rust}
\title{Source rust file}
\usage{
rust(path, code, depend = NULL, rebuild = FALSE)
}
\arguments{
\item{path}{package path}

\item{code}{rust code}

\item{depend}{rustr version}

\item{rebuild}{rebuild cache}
}
\description{
Source rust file
}

