% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust.R
\docType{data}
\name{RUSTINR_TEMP}
\alias{RUSTINR_TEMP}
\title{Temporary folder for \code{rust()} function}
\format{An object of class \code{environment} of length 1.}
\usage{
RUSTINR_TEMP
}
\description{
\code{rust()} function creates a temporary folder for Rust code generation. Users can get the path to this folder with \code{RUSTINR_TEMP$obj}, and get the ramdon string which is the prefix of generated Rust functions with \code{RUSTINR_TEMP$rss}.
}
\examples{
\dontrun{
check_rustr()
RUSTINR_TEMP$obj
list.dirs(RUSTINR_TEMP$obj)
RUSTINR_TEMP$rss

}
}
\keyword{datasets}

