% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustr_check.R
\name{rustr_check}
\alias{rustr_check}
\title{Check rustinr status}
\usage{
rustr_check(detail = FALSE)
}
\arguments{
\item{detail}{print detail info, FALSE by default.}
}
\description{
\code{rustr_check} checks the status of rustinr, rustc, and cargo installation.
}

