% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_build.R
\name{rustr_clean_build}
\alias{rustr_clean_build}
\title{Clean Rust build target}
\usage{
rustr_clean_build(pkgdir = ".", lib_only = FALSE)
}
\arguments{
\item{pkgdir}{package path}

\item{lib_only}{only remove builded static library}
}
\description{
\code{rustr_clean_build} cleans the build target of \code{cargo build}.
}

