% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{rustr_init}
\alias{rustr_init}
\title{Create a Skeleton for a New Source Package}
\usage{
rustr_init(name, path = ".", force = FALSE, code_files = character(),
  rust_files = character(), author = "Your Name", maintainer = if
  (missing(author)) "Your Name" else author, email = "your@email.com",
  license = "MIT")
}
\arguments{
\item{name}{character string: the package name and directory name for your package.}

\item{path}{path to put the package directory in.}

\item{force}{If FALSE will not overwrite an existing directory.}

\item{code_files}{a character vector with the paths to R code files to build the package around.}

\item{rust_files}{rust files}

\item{author}{author}

\item{maintainer}{maintainer}

\item{email}{email}

\item{license}{license}
}
\description{
Create a Skeleton for a New Source Package
}

