% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust.R
\docType{data}
\name{RUSTR_TEMP}
\alias{RUSTR_TEMP}
\title{Temporary folder for \code{rust()} function}
\format{An object of class \code{environment} of length 1.}
\usage{
RUSTR_TEMP
}
\description{
\code{rust()} function creates a temporary folder for Rust code generation. Users can get the path to this folder with \code{RUSTR_TEMP$obj}, and get the ramdon string which is the prefix of generated Rust functions with \code{RUSTR_TEMP$rss}.
}
\examples{
\dontrun{

rustr_check()

RUSTR_TEMP$obj

list.dirs(RUSTR_TEMP$obj)

RUSTR_TEMP$rss

readLines(file.path(RUSTR_TEMP$obj, "src", "REXPORT.c"))
}
}
\keyword{datasets}

