% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\docType{data}
\name{SOURCE_RUST_PATH}
\alias{SOURCE_RUST_PATH}
\title{Temporary folder for source_rust()}
\format{An object of class \code{environment} of length 1.}
\usage{
SOURCE_RUST_PATH
}
\description{
Temporary folder for source_rust()
}
\keyword{datasets}

