% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rustr.R
\name{check_rustr}
\alias{check_rustr}
\title{Check rustr installation}
\usage{
check_rustr(detail = FALSE)
}
\arguments{
\item{detail}{show more detail}
}
\description{
Check rustr installation
}

