% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rustr.R
\name{check_rustr}
\alias{check_rustr}
\title{Check that rustr is installed}
\usage{
check_rustr(detail = FALSE)
}
\arguments{
\item{detail}{Whether to provide more detail about the success or failure of
the check. Set to FALSE by default.}
}
\description{
rust/R integration depends on both the \code{rustinr} R package
and then \code{rustr} rust crate. This function tests that \code{rustr} is
installed.
}

