% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rust}
\alias{rust}
\title{Source rust file}
\usage{
rust(code, path, depend = NULL, rebuild = FALSE, env = globalenv())
}
\arguments{
\item{code}{rust code}

\item{path}{package path}

\item{depend}{rustr version}

\item{rebuild}{rebuild cache}

\item{env}{an environment, determining where the export R functions are evaluated}
}
\description{
Source rust file
}

