% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rust}
\alias{rust}
\title{Source Rust file}
\usage{
rust(code, path, depend = NULL, header = TRUE, rebuild = FALSE,
  env = globalenv())
}
\arguments{
\item{code}{Rust code}

\item{path}{Rust file path}

\item{depend}{Rust dependencies}

\item{header}{Add default Rust header}

\item{rebuild}{Force rebuild cache}

\item{env}{An environment, determining where the export R functions are evaluated}
}
\description{
Source Rust file
}

