% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{rustr_init}
\alias{rustr_init}
\title{Create a skeleton for a new source package}
\usage{
rustr_init(name, path = ".", force = FALSE, code_files = character(),
  rust_files = character(), author = "Your Name", maintainer = if
  (missing(author)) "Your Name" else author, email = "your@email.com",
  license = "MIT")
}
\arguments{
\item{name}{A character string: the package name and directory name for your package.}

\item{path}{A path to put the package directory in.}

\item{force}{See \link[utils]{package.skeleton}}

\item{code_files}{See \link[utils]{package.skeleton}}

\item{rust_files}{A character vector with the paths to Rust source files to add to the package.}

\item{author}{Author of the package.}

\item{maintainer}{Maintainer of the package.}

\item{email}{Email of the package maintainer.}

\item{license}{License of the package.}
}
\description{
Create a skeleton for a new source package
}
\references{
Read the \emph{Writing R Extensions} manual for more details.
}

