% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{rustr_init}
\alias{rustr_init}
\title{Create a skeleton for a new source package with Rust support}
\usage{
rustr_init(name, path = ".", force = FALSE, code_files = character(),
  rust_files = character(), author = "Your Name", maintainer = if
  (missing(author)) "Your Name" else author, email = "your@email.com",
  license = "MIT")
}
\arguments{
\item{name}{A character string: the package name and directory name for your package.}

\item{path}{A path to put the package directory in.}

\item{force}{See \link[utils]{package.skeleton}}

\item{code_files}{See \link[utils]{package.skeleton}}

\item{rust_files}{A character vector with the paths to Rust source files to add to the package.}

\item{author}{Author of the package.}

\item{maintainer}{Maintainer of the package.}

\item{email}{Email of the package maintainer.}

\item{license}{License of the package.}
}
\description{
Create a skeleton for a new source package with Rust support.
}
\details{
There is a folder name \code{rustlib} in the created package \code{src} path. This folder is a rust library which will be compiled and linked by R. Just write Rust code in \code{./src/rustlib/src/} .

\code{./src/rustlib/src/export.rs}, \code{./src/REXPORT.c} and \code{./R/REXPORT.R} are generated by \code{rustrize()}, and they should not be edited by hand. Before building the package, run \code{rustrize()} to generate these three files.
}
\examples{
\dontrun{

rustr_init("test-pkg")
}
}
\references{
Read the \emph{Writing R Extensions} manual for more details.
}
\seealso{
https://book.rustr.org/create-a-rust-r-package.html
}

